Reorganize crate
The addition of the web gem module means that main is too big.
- Break extern functions into ffi module.
- Break interpreter state and wrapper into interpreter module.
